rm models.ts
