+++ src/contexts/use-chat/chat-context.tsx
@@ -16,7 +16,7 @@ import { useModel } from "@/contexts/use-model/model-hooks";
 import { usePersistence } from "@/contexts/use-persistence/persistence-hooks";
 import { useChat } from "@/hooks/ai/use-chat";
 import { useModelCatalog } from "@/hooks/ai/use-model-catalog";
-import { type ModelConfig, type ModelData } from "@/lib/ai/models";
+import { type ModelConfig, type ModelData } from "@/hooks/ai/use-model-catalog";
 import { chatIdsAtom } from "@/lib/jotai/atoms";
 import { notificationSettingAtom } from "@/lib/jotai/settings-atoms";
 import { getLogger } from "@/lib/logger";
